Relocate task-contract / feedback / policy helpers to actor_loop_flow
Move 12 small pure-fn helpers (88 LOC of function body, 129 LOC
file-level net) out of turn.rs into actor_loop_flow.rs where their
production callers already live. No behavioural change — pure code
motion plus orphaned doc-comment cleanup.

Moved helpers (all already pub(super)):

  Feedback factories (src/session::feedback):
    - build_feedback_for_unsafe_block
    - build_feedback_for_no_repo_progress
    - build_feedback_for_no_tool_call

  Recovery / status notes:
    - should_record_no_repo_progress
    - missing_repo_change_retry_status_note
    - rejected_tool_batch_retry_status_note
    - focused_policy_retry_exhausted
    - unrestricted_policy_retry_exhausted
    - should_try_framework_app_fallback
    - framework_app_fallback_continuation_note

  Task-contract notes / safe-stop mapping:
    - task_contract_safe_stop_clear_tag
    - task_contract_verifier_edit_required_note

Required adjustments:

  * actor_loop_flow.rs: added FeedbackFrame / FeedbackFrameDraft /
    build_feedback_frame to the existing session::feedback import,
    dropped the 5 super::turn::* entries that are now local.

  * turn.rs: ActorLoopMissingRepoChangeReplyKind dropped from the
    top-of-file actor_loop_flow re-import (only the test mod uses it
    now, via its own targeted `use`); 3 use entries in mod
    progress_tests / mod tests moved into a dedicated
    `use super::super::actor_loop_flow::{...}` for the relocated fns;
    8 `super::FN(` test call sites rewritten to
    `super::super::actor_loop_flow::FN(`.

  * Orphaned doc-comment blocks above the relocated functions
    removed (clippy::empty_line_after_doc_comments fail-closed under
    `-D warnings`).

Metrics: turn.rs 35,451 -> 35,322 (-129); actor_loop_flow.rs
4,260 -> 4,359 (+99). Remaining gap to Issue #681 acceptance
(turn.rs <= 35,000): ~322 LOC.
